moved business logic in services directory
