+++ package.json
@@ -87,7 +87,7 @@
     "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.8",
     "@tanstack/react-form": "^1.29.1",
-    "@tanstack/react-query": "^5.100.1",
+    "@tanstack/react-query": "^5.100.5",
     "@tanstack/react-router": "1.168.23",
     "@tanstack/react-router-devtools": "1.166.13",
     "@tanstack/react-start": "1.167.43",
@@ -115,7 +115,7 @@
     "react": "^19.2.5",
     "react-day-picker": "9.14.0",
     "react-dom": "^19.2.5",
-    "react-hook-form": "^7.73.1",
+    "react-hook-form": "^7.74.0",
     "react-markdown": "^10.1.0",
     "react-top-loading-bar": "^3.0.2",
     "read-pkg": "^10.1.0",
@@ -139,7 +139,7 @@
     "@playwright/test": "^1.59.1",
     "@rsbuild/core": "^2.0.1",
     "@tailwindcss/vite": "^4.2.4",
-    "@tanstack/react-query-devtools": "^5.100.1",
+    "@tanstack/react-query-devtools": "^5.100.5",
     "@testing-library/react": "^16.3.2",
     "@types/bun": "^1.3.13",
     "@types/hast": "^3.0.4",
@@ -157,7 +157,7 @@
     "fdir": "^6.5.0",
     "jsdom": "^29.0.2",
     "k6": "^0.0.0",
-    "knip": "^6.6.3",
+    "knip": "^6.7.0",
     "oxfmt": "^0.42.0",
     "oxlint": "^1.61.0",
     "portless": "^0.10.3",
