+++ src/features/auth/sign-up/components/sign-up-form.tsx
@@ -12,9 +12,9 @@ import { useNavigate } from "@tanstack/react-router";
 import { UserPlus, Loader2 } from "lucide-react";
 import { toast } from "sonner";
 import { authClient, signUpWithEmail } from "~/lib/auth/client";
-import { authActions } from "~/lib/stores/auth-store";
+import { authActions } from "~/lib/stores/auth";
 import { env } from "~/config/env";
-import { getEnabledSocialProviders, type AuthProviderId } from "~/config/auth-providers";
+import { getEnabledSocialProviders, type AuthProviderId } from "~/config/auth";
 import { cn } from "~/lib/utils";
 import { encodePassword } from "~/lib/utils/encryption";
 import { Button } from "~/components/ui/button";
