+++ src/features/settings/notifications/index.tsx
@@ -1,6 +1,6 @@
 import { ContentSection } from "../components/content-section";
 import { NotificationsForm } from "./notifications-form";
-import { NotificationData } from "~/lib/stores/settings-store";
+import { NotificationData } from "~/lib/stores/settings";
 
 /**
  * Notifications settings page component.
