+++ src/features/settings/profile/profile-form.tsx
@@ -18,7 +18,7 @@ import {
 } from "~/components/ui/form";
 import { Input } from "~/components/ui/input";
 import { Textarea } from "~/components/ui/textarea";
-import { settingsActions } from "~/lib/stores/settings-store";
+import { settingsActions } from "~/lib/stores/settings";
 import { type ProfileData } from ".";
 
 /**
