+++ src/routes/__root.tsx
@@ -13,7 +13,7 @@ import { DirectionProvider } from "~/context/direction-provider";
 import { FontProvider } from "~/context/font-provider";
 import { ThemeProvider } from "~/context/theme-provider";
 import { useAuthSync } from "~/lib/auth/store-sync";
-import { useAuthInit } from "~/lib/stores/auth-store";
+import { useAuthInit } from "~/lib/stores/auth";
 import { queryClient } from "~/router";
 import appCss from "~/styles/app.css?url";
 import { SearchProvider } from "~/context/search-provider";
